% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Value matching for absences}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.}

\item{table}{vector or \code{NULL}: the values to be matched against}
}
\description{
This provides a more succinct and intuitive interface to determine if a value doesn't exist in a table.
Avoids difficult to type and read \code{!a \%in\% b}
}
\examples{
\dontrun{
 filter(mtcars, carb \%notin\% c(1,3)) }
}
